Fixed mosaic and avgpbz script calls and JAWs_products check.
The mos.py and avgpbz.py scripts were broken because of calling being a leftover from the Southampton machine setup. Have changed the scripts calls to the same format as rsmppfuncs. Should now work on any machine.

Also added is a check for JAWS_products existence before attempting to create it.

All edits are for lofar-imager.py.